refactor(variance): separate alert management service
Alert rule management and incident lifecycle handling were sharing the variance facade with calculation and coordinator code. Move that alert-specific behavior into a focused variance-tracking submodule while preserving the public facade export and constructor wiring.

Constraint: Routes, automation, and tests continue importing AlertManagementService through server/services/variance-tracking

Rejected: Move VarianceCalculationService in this commit | calculation logic is the largest remaining surface and needs its own review

Confidence: high

Scope-risk: moderate

Directive: Do not import from the facade inside variance-tracking submodules; submodules should depend on helpers or sibling modules directly

Tested: npm run check

Tested: npx vitest run --config vitest.config.mjs --configLoader native --project=server tests/unit/services/variance-tracking.test.ts tests/unit/api/variance-tracking-api.test.ts tests/unit/services/variance-alert-automation.test.ts

Not-tested: Full npm test suite
